application/requirements.txt
djangorestframework
Django
psycopg2-binary
knox
cryptography
Jinja2
boto3
validate_email
django-heroku
gunicorn
